+++ extension/content.js
@@ -1,0 +1,14 @@
+// content.js
+
+// Example content script that interacts with the web page content
+// For illustration purposes only, replace with your actual content script logic
+
+// Example: Listen for page load event
+document.addEventListener('DOMContentLoaded', function() {
+    console.log('Page loaded');
+
+    // Example: Send a message to the background script
+    chrome.runtime.sendMessage({ greeting: 'Hello from content script!' }, function(response) {
+        console.log('Response from background script:', response);
+    });
+});
