+++ extension/manifest.json
@@ -1,0 +1,29 @@
+{
+  "manifest_version": 3,
+  "name": "Fake Detector",
+  "version": "1.0",
+  "description": "Detects fake images and videos",
+  "permissions": ["activeTab", "storage", "webNavigation"],
+  "background": {
+    "service_worker": "background.js"
+  },
+  "content_scripts": [
+    {
+      "matches": ["<all_urls>"],
+      "js": ["content.js"]
+    }
+  ],
+  "action": {
+    "default_popup": "popup.html",
+    "default_icon": {
+      "16": "icons/icon16.png",
+      "48": "icons/icon48.png",
+      "128": "icons/icon128.png"
+    }
+  },
+  "icons": {
+    "16": "icons/icon16.png",
+    "48": "icons/icon48.png",
+    "128": "icons/icon128.png"
+  }
+}
